promex v0.0.1 Promex.Gauge

Gauge keeps metrics that can go up and down.

Functions

dec(name)

See Promex.Gauge.decrement/1.

dec(name, opts)

See Promex.Gauge.decrement/2.

decrement(name)

Decrement the gauge identified by name by 1, or initialize it with a value of -1.

iex> Promex.Gauge.decrement("test_gauge_4")
{:ok, -1}
iex> Promex.Gauge.decrement("test_gauge_4")
{:ok, -2}

decrement(name, list)

Decrement the gauge identified by name by a specific value.

By default, this will set the gauge value to -1. To set a different initial value, use the by option.

iex> Promex.Gauge.decrement("test_gauge_5", by: 5)
{:ok, -5}
iex> Promex.Gauge.decrement("test_gauge_5", by: 5)
{:ok, -10}

inc(name)

See Promex.Gauge.increment/1.

inc(name, opts)

See Promex.Gauge.increment/2.

increment(name)

Increment the gauge identified by name by 1, or initialize it with a value of 1.

iex> Promex.Gauge.increment("test_gauge_2")
{:ok, 1}
iex> Promex.Gauge.increment("test_gauge_2")
{:ok, 2}

increment(name, list)

Increment the gauge identified by name by a specific value.

iex> Promex.Gauge.set("test_gauge_3")
{:ok, 1}
iex> Promex.Gauge.increment("test_gauge_3", by: 5)
{:ok, 6}

set(name)

Set a gauge’s value.

iex> Promex.Gauge.set("test_gauge_1")
{:ok, 1}
iex> Promex.Gauge.set("test_gauge_1", 10)
{:ok, 10}

set(name, value)